Rename search component
